fix: サーボ angle() を 50Hz 標準化 + ファーム MUL_Q16 の int16 乗算バグ修正
- angle() を 50Hz・パルス 0.5〜2.4ms に変更（Arduino Servo とほぼ同等レンジ、
  分解能約7°/ステップ）。servo.frequency(50) 必須ガード付き。
  旧 500Hz 設計はパルス範囲が狭い上、大角度で Low ギャップが消滅して
  サーボがパルス境界を認識できなかった
- ファーム MUL_Q16 の int16 キャストを除去（int32 乗算、積 ±32768 未満まで正確）。
  ±128 以上のオペランドを含む乗算すべてが壊れる一般バグだった。
  Q1 系 4 スケッチ再生成（他 10 本はバイト一致）— 書き込み済みボードは要再書き込み
- angle 変数経路は「÷180 → ×24.32」の順で計算し旧ファームでも正しく動く
- Flash: Q1 +596 → +592B。Q1Sv 16212B（98%）。COMP_FLASH・バッジ・命令セット注記更新
- サンプル pwm-servo / pwm-servo-q1・API 表を frequency(50) に更新
- tools/gen-ino.js の旧ファイル名参照（urb-lab.html）を修復
- SG90 実機で 0〜180° 全域動作を確認済み。全36サンプル回帰 OK

+++ docs/sketches/UIAPrubyVmQ1Sv/UIAPrubyVmQ1Sv.ino
@@ -432,7 +432,8 @@ static bool runUap(const char *filename) {
       case OP_SUB_Q16: { uint8_t b[2]; if(sm_read_full(b,2)!=2) goto vm_err; pc+=2; regs[b[0]&3]-=regs[b[1]&3]; break; }
       case OP_MUL_Q16: {
         uint8_t b[2]; if(sm_read_full(b,2)!=2) goto vm_err; pc+=2;
-        regs[b[0]&3] = ((int32_t)(int16_t)regs[b[0]&3] * (int16_t)regs[b[1]&3]) >> 8; break;
+        // int32 乗算（旧 int16 キャストは ±128 以上のオペランドを壊すバグだった）。積が ±32768 未満まで正確
+        regs[b[0]&3] = (regs[b[0]&3] * regs[b[1]&3]) >> 8; break;
       }
       case OP_DIV_Q16: {
         uint8_t b[2]; if(sm_read_full(b,2)!=2) goto vm_err; pc+=2;
